fix(columnar): add lossless i64 block stats for integer predicate pushdown
f64 cannot represent all i64 values exactly (values outside ±2^53 are
rounded), causing block-skip predicates on large integer or timestamp
columns to produce incorrect results — blocks that should be skipped
are read, and in edge cases blocks that must not be skipped are skipped.

Fix by tracking exact integer bounds alongside the existing f64 stats:

- `BlockStats` gains optional `min_i64`/`max_i64` fields
  (`#[serde(default)]` for backward compatibility with minor v0 segments).
- `BlockStats::integer(min, max, ...)` constructor populates both the
  lossless i64 fields and the lossy f64 fields so f64-based paths remain
  a valid conservative fallback.
- `format.rs` bumps `VERSION_MINOR` to 1.
- `PredicateValue::Integer(i64)` variant added; `Predicate` gains
  `eq_i64`, `ne_i64`, `gt_i64`, `gte_i64`, `lt_i64`, `lte_i64`
  constructors.
- `can_skip_integer`: when `min_i64`/`max_i64` are present comparison is
  done entirely in i64; for segments written at minor v0, falls through to
  a conservative f64 path (may fail to skip, never incorrect).
- `writer/block.rs`: i64 and timestamp columns call `BlockStats::integer`.
- `writer/mod.rs` and `timeseries/columnar_bridge.rs`: planner emits
  `PredicateValue::Integer` for integer predicate operands.

+++ nodedb-columnar/src/writer/block.rs
@@ -63,8 +63,7 @@ fn encode_single_block(
             let null_count = valid_slice.iter().filter(|&&v| !v).count() as u32;
 
             let (min, max) = numeric_min_max_i64(slice, valid_slice);
-            let stats =
-                BlockStats::numeric(min as f64, max as f64, null_count, block_row_count as u32);
+            let stats = BlockStats::integer(min, max, null_count, block_row_count as u32);
 
             let encoded = encode_i64_with_validity(slice, valid_slice, codec)?;
             Ok((encoded, stats))
@@ -86,8 +85,7 @@ fn encode_single_block(
             let null_count = valid_slice.iter().filter(|&&v| !v).count() as u32;
 
             let (min, max) = numeric_min_max_i64(slice, valid_slice);
-            let stats =
-                BlockStats::numeric(min as f64, max as f64, null_count, block_row_count as u32);
+            let stats = BlockStats::integer(min, max, null_count, block_row_count as u32);
 
             let encoded = encode_i64_with_validity(slice, valid_slice, codec)?;
             Ok((encoded, stats))
